move mock router into own file
